chore: move workspace apps tests to new package
